backend/requirements.txt
# ── Parseo de documentos ───────────────────────────────────────────────────
docling==2.85.0
docling-core==2.73.0

# ── RAG / Chunking ─────────────────────────────────────────────────────────
llama-index-core==0.14.20
tiktoken==0.12.0

# ── LLM y embeddings ───────────────────────────────────────────────────────
openai>=1.30.0                      # GPT-4o vision + text-embedding-3-small

# ── Vector store ───────────────────────────────────────────────────────────
chromadb>=0.5.0                     # dev local
# azure-search-documents>=11.4.0   # producción Azure AI Search

# ── API ────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── Config ─────────────────────────────────────────────────────────────────
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
